Split test utilities into more relevant files.
