+++ action_test.go
@@ -1,0 +1,17 @@
+package testkit_test
+
+import (
+	"context"
+
+	. "github.com/dogmatiq/testkit"
+)
+
+// noop is an Action that does nothing.
+var noop noopAction
+
+// noopAction is an Action that does nothing. It is intended to be used for
+// testing the test system itself.
+type noopAction struct{}
+
+func (noopAction) ExpectOptions() []ExpectOption                  { return nil }
+func (noopAction) Apply(ctx context.Context, s ActionScope) error { return nil }
